Moved the Network files back
Since other mods seem to depend on them being in that specific location.
