+++ src/main/java/baubles/common/event/KeyHandler.java
@@ -3,7 +3,7 @@ package baubles.common.event;
 import org.lwjgl.input.Keyboard;
 
 import baubles.common.network.PacketHandler;
-import baubles.common.network.server.PacketOpenBaublesInventory;
+import baubles.common.network.PacketOpenBaublesInventory;
 import cpw.mods.fml.client.registry.ClientRegistry;
 import cpw.mods.fml.common.gameevent.InputEvent;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
