+++ src/main/java/baubles/common/network/PacketOpenNormalInventory.java
@@ -1,4 +1,4 @@
-package baubles.common.network.server;
+package baubles.common.network;
 
 import cpw.mods.fml.common.network.simpleimpl.IMessage;
 import cpw.mods.fml.common.network.simpleimpl.IMessageHandler;
